+++ noodles-bgzf/src/io/multithreaded_writer/builder.rs
@@ -38,6 +38,10 @@ impl Builder {
     /// let builder = bgzf::io::multithreaded_writer::Builder::default()
     ///     .set_worker_count(NonZero::<usize>::MIN);
     /// ```
+    #[deprecated(
+        since = "0.48.0",
+        note = "Use `rayon::ThreadPoolBuilder` to configure the thread pool."
+    )]
     pub fn set_worker_count(mut self, worker_count: NonZero<usize>) -> Self {
         self.worker_count = worker_count;
         self
@@ -60,7 +64,7 @@ impl Builder {
     {
         use super::{State, spawn_writer};
 
-        let worker_count = self.worker_count.get();
+        let worker_count = rayon::current_num_threads();
 
         let (write_tx, write_rx) = crossbeam_channel::bounded(worker_count);