+++ noodles-csi/src/io/writer/index/header.rs
@@ -3,7 +3,10 @@ mod reference_sequence_names;
 use std::io::{self, Write};
 
 use self::reference_sequence_names::write_reference_sequence_names;
-use crate::{binning_index::index::Header, io::writer::num::write_i32_le};
+use crate::{
+    binning_index::index::{Header, header::Format},
+    io::writer::num::write_i32_le,
+};
 
 pub(super) fn write_aux<W>(writer: &mut W, header: Option<&Header>) -> io::Result<()>
 where
@@ -28,39 +31,99 @@ pub(crate) fn write_header<W>(writer: &mut W, header: &Header) -> io::Result<()>
 where
     W: Write,
 {
-    let format = i32::from(header.format());
-    write_i32_le(writer, format)?;
+    write_format(writer, header.format())?;
+    write_reference_sequence_name_index(writer, header.reference_sequence_name_index())?;
+    write_start_position_index(writer, header.start_position_index())?;
 
-    let reference_sequence_name_index = header
-        .reference_sequence_name_index()
-        .checked_add(1)
-        .expect("attempt to add with overflow");
-    let col_seq = i32::try_from(reference_sequence_name_index)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
-    write_i32_le(writer, col_seq)?;
+    write_end_position_index(
+        writer,
+        header.format(),
+        header.start_position_index(),
+        header.end_position_index(),
+    )?;
 
-    let start_position_index = header
-        .start_position_index()
-        .checked_add(1)
-        .expect("attempt to add with overflow");
-    let col_beg = i32::try_from(start_position_index)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
-    write_i32_le(writer, col_beg)?;
-
-    let col_end = header.end_position_index().map_or(Ok(0), |mut i| {
-        i = i.checked_add(1).expect("attempt to add with overflow");
-        i32::try_from(i).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))
-    })?;
-    write_i32_le(writer, col_end)?;
-
-    let meta = i32::from(header.line_comment_prefix());
-    write_i32_le(writer, meta)?;
-
-    let skip = i32::try_from(header.line_skip_count())
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
-    write_i32_le(writer, skip)?;
+    write_line_comment_prefix(writer, header.line_comment_prefix())?;
+    write_line_skip_count(writer, header.line_skip_count())?;
 
     write_reference_sequence_names(writer, header.reference_sequence_names())?;
+
+    Ok(())
+}
+
+fn write_end_position_index<W>(
+    writer: &mut W,
+    format: Format,
+    start_position_index: usize,
+    end_position_index: Option<usize>,
+) -> io::Result<()>
+where
+    W: Write,
+{
+    const SPECIALIZED_END_VALUE: i32 = 0;
+
+    if matches!(format, Format::Sam | Format::Vcf) {
+        if end_position_index.is_some() {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidInput,
+                "invalid end position index for format",
+            ));
+        } else {
+            write_i32_le(writer, SPECIALIZED_END_VALUE)?;
+        }
+    } else {
+        let i = end_position_index.unwrap_or(start_position_index);
+        let j = i.checked_add(1).expect("attempt to add with overflow");
+        let n = i32::try_from(j).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+        write_i32_le(writer, n)?;
+    }
+
+    Ok(())
+}
+
+fn write_format<W>(writer: &mut W, format: Format) -> io::Result<()>
+where
+    W: Write,
+{
+    let n = i32::from(format);
+    write_i32_le(writer, n)
+}
+
+fn write_reference_sequence_name_index<W>(writer: &mut W, i: usize) -> io::Result<()>
+where
+    W: Write,
+{
+    let j = i.checked_add(1).expect("attempt to add with overflow");
+    let n = i32::try_from(j).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+    write_i32_le(writer, n)?;
+    Ok(())
+}
+
+fn write_start_position_index<W>(writer: &mut W, i: usize) -> io::Result<()>
+where
+    W: Write,
+{
+    let j = i.checked_add(1).expect("attempt to add with overflow");
+    let n = i32::try_from(j).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+    write_i32_le(writer, n)?;
+    Ok(())
+}
+
+fn write_line_comment_prefix<W>(writer: &mut W, line_comment_prefix: u8) -> io::Result<()>
+where
+    W: Write,
+{
+    let n = i32::from(line_comment_prefix);
+    write_i32_le(writer, n)
+}
+
+fn write_line_skip_count<W>(writer: &mut W, line_skip_count: u32) -> io::Result<()>
+where
+    W: Write,
+{
+    let n = i32::try_from(line_skip_count)
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+
+    write_i32_le(writer, n)?;
 
     Ok(())
 }
@@ -70,6 +133,7 @@ mod tests {
     use bstr::BString;
 
     use super::*;
+    use crate::binning_index::index::header::format::CoordinateSystem;
 
     #[test]
     fn test_write_aux() -> io::Result<()> {
@@ -102,6 +166,49 @@ mod tests {
         ];
 
         assert_eq!(buf, expected);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_write_end_position_index() -> io::Result<()> {
+        fn t(
+            buf: &mut Vec<u8>,
+            format: Format,
+            start_position_index: usize,
+            end_position_index: Option<usize>,
+            expected: i32,
+        ) -> io::Result<()> {
+            buf.clear();
+            write_end_position_index(buf, format, start_position_index, end_position_index)?;
+            assert_eq!(buf, &expected.to_le_bytes());
+            Ok(())
+        }
+
+        let mut buf = Vec::new();
+
+        t(&mut buf, Format::Sam, 5, None, 0)?;
+        t(&mut buf, Format::Vcf, 5, None, 0)?;
+        t(&mut buf, Format::Generic(CoordinateSystem::Gff), 5, None, 6)?;
+        t(
+            &mut buf,
+            Format::Generic(CoordinateSystem::Gff),
+            5,
+            Some(8),
+            9,
+        )?;
+
+        buf.clear();
+        assert!(matches!(
+            write_end_position_index(&mut buf, Format::Sam, 5, Some(8)),
+            Err(e) if e.kind() == io::ErrorKind::InvalidInput
+        ));
+
+        buf.clear();
+        assert!(matches!(
+            write_end_position_index(&mut buf, Format::Vcf, 5, Some(8)),
+            Err(e) if e.kind() == io::ErrorKind::InvalidInput
+        ));
 
         Ok(())
     }